✅ (dashboard): Update smoke tests to wrap with AuthProvider
