+++ dashboard/src/App.test.tsx
@@ -1,10 +1,11 @@
 import { render, screen } from '@testing-library/react'
-import { MemoryRouter } from 'react-router-dom'
+import { MemoryRouter, Routes, Route } from 'react-router-dom'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { AuthProvider } from './auth/AuthProvider'
+
 import { ProtectedRoute } from './pages/ProtectedRoute'
 import { LoginPage } from './pages/LoginPage'
 import { NotFoundPage } from './pages/NotFoundPage'
-import { Routes, Route } from 'react-router-dom'
 
 function makeClient() {
   return new QueryClient({ defaultOptions: { queries: { retry: false } } })
@@ -14,13 +15,15 @@ function AppRoutes({ initialPath = '/' }: { initialPath?: string }) {
   return (
     <QueryClientProvider client={makeClient()}>
       <MemoryRouter initialEntries={[initialPath]}>
-        <Routes>
-          <Route path="/login" element={<LoginPage />} />
-          <Route element={<ProtectedRoute />}>
-            <Route path="/" element={<div>Dashboard home</div>} />
-          </Route>
-          <Route path="*" element={<NotFoundPage />} />
-        </Routes>
+        <AuthProvider>
+          <Routes>
+            <Route path="/login" element={<LoginPage />} />
+            <Route element={<ProtectedRoute />}>
+              <Route path="/" element={<div>Dashboard home</div>} />
+            </Route>
+            <Route path="*" element={<NotFoundPage />} />
+          </Routes>
+        </AuthProvider>
       </MemoryRouter>
     </QueryClientProvider>
   )
@@ -39,7 +42,7 @@ describe('Router smoke tests', () => {
   it('renders LoginPage at /login', () => {
     render(<AppRoutes initialPath="/login" />)
     expect(screen.getByRole('heading', { name: 'Agent Assembly' })).toBeInTheDocument()
-    expect(screen.getByLabelText('API Token')).toBeInTheDocument()
+    expect(screen.getByLabelText('API Key')).toBeInTheDocument()
   })
 
   it('renders NotFoundPage for unknown routes', () => {
