+++ dashboard/src/pages/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom'
-import { useAuth } from '../auth/AuthProvider'
+import { useAuth } from '../auth/useAuth'
 
 export function ProtectedRoute() {
   const { token } = useAuth()
